LPS-166190: Move SentenceEmbeddingRetriever to API module
