+++ modules/dxp/apps/search-experiences/search-experiences-api/src/main/resources/com/liferay/search/experiences/ml/sentence/embedding/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
